pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "grandalf>=0.8",
    "isort>=6.1.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.8",
    "python-dotenv>=1.1.1",
    "tweepy>=4.14.0",
]
